Renamed config to constants
