+++ tests/integration/test_lr_dynamic_pharmacophore.py
@@ -34,5 +34,5 @@ def test_dynamic_ligand_receptor_pharmacophore(traj_er_alpha):
 
     # Finally we visualize the pharmacophore.
     viewer = oph.Viewer()
-    viewer.add_components([protein, ligand])
-    viewer.show()
+    viewer.add_components([protein, ligand, pharmacophore])
+    viewer.show(frame=1)
